+++ functions/src/params.ts
@@ -1,0 +1,4 @@
+import { defineString } from "firebase-functions/params"
+
+export const STRIPE_PRICE_API_ID$ = defineString("STRIPE_PRICE_API_ID")
+export const HOSTING_ORIGIN$ = defineString("HOSTING_ORIGIN")
